Rename: 신고 관련 컴포넌트 위치 이동
