+++ src/components/Report/Comment/ReportComment.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import Pagination from '../Pagination/Pagination';
+import Pagination from '../../Pagination/Pagination';
 import { useQuery } from 'react-query';
-import { getReportedCommentListRequest } from '../../apis/adminApi';
+import { getReportedCommentListRequest } from '../../../apis/adminApi';
 import { useSearchParams } from 'react-router-dom';
 
-import ReportCommentTableRow from '../TableRow/ReportCommentTableRow';
-import MessageBox from '../MessageBox';
+import ReportCommentTableRow from './ReportCommentTableRow';
+import MessageBox from '../../MessageBox';
 
 function ReportComment() {
   const [searchParams] = useSearchParams();
@@ -22,6 +22,7 @@ function ReportComment() {
 
   const reportComments = data?.data.data.reportComments;
   const isEmpty = !reportComments?.length;
+
   if (isEmpty) return <MessageBox msg={`😊 신고내역이 없습니다!`} />;
   return (
     <div className="flex flex-col gap-2 ">
@@ -39,7 +40,7 @@ function ReportComment() {
         );
       })}
       <Pagination
-          pageCount={data?.data.data.totalPageSize || 0}
+        pageCount={data?.data.data.totalPageSize || 0}
         listLength={data?.data.data.size || 0}
       />
     </div>
